# Speed up find_neighbor_sites and num_neighbors