--- a/examples/tutorials/Intro to OpenPNM - Intermediate.ipynb
+++ b/examples/tutorials/Intro to OpenPNM - Intermediate.ipynb
@@ -55,14 +55,14 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "pn = op.network.Cubic(shape=[20, 20, 10], spacing=0.0001, connectivity=8)"
+    "pn = op.network.Cubic(shape=[20, 20, 10], spacing=0.0001, connectivity=14)"
    ]
   },
   {
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-    "* This **Network** has pores distributed in a cubic lattice, but connected to diagonal neighbors due to the ``connectivity`` being set to 8 (the default is 6 which is orthogonal neighbors).  The various options are outlined in the **Cubic** class's documentation which can be viewed with the Object Inspector in Spyder.\n",
+    "* This **Network** has pores distributed in a cubic lattice, but connected to diagonal neighbors due to the ``connectivity`` being set to 14 (the default is 6 which is orthogonal neighbors).  The various options are outlined in the **Cubic** class's documentation which can be viewed with the Object Inspector in Spyder.\n",
     "\n",
     "* OpenPNM includes several other classes for generating networks including random topology based on Delaunay tessellations (**Delaunay**).\n",
     "\n",
@@ -266,9 +266,9 @@
     {
      "data": {
       "text/plain": [
-       "array([[5.21639565e-05, 5.88927751e-05],\n",
-       "       [5.83316363e-05, 5.43440496e-05],\n",
-       "       [5.63074931e-05, 5.81044386e-05],\n",
+       "array([[5.21639565e-05, 5.83316363e-05],\n",
+       "       [5.83316363e-05, 5.63074931e-05],\n",
+       "       [5.63074931e-05, 5.92495511e-05],\n",
        "       ...,\n",
        "       [5.95407077e-05, 5.39710315e-05],\n",
        "       [5.31619482e-05, 5.42961235e-05],\n",
@@ -803,7 +803,7 @@
      "name": "stdout",
      "output_type": "stream",
      "text": [
-      "[1.43421902e-10] [3.17745347e-10] [1.13961253e-10]\n"
+      "[1.8792235e-10] [4.46916277e-10] [1.69341333e-10]\n"
      ]
     }
    ],
@@ -832,7 +832,7 @@
    "name": "python",
    "nbconvert_exporter": "python",
    "pygments_lexer": "ipython3",
-   "version": "3.8.1"
+   "version": "3.8.3"
   }
  },
  "nbformat": 4,