--- a/openpnm/network/GenericNetwork.py
+++ b/openpnm/network/GenericNetwork.py
@@ -9,6 +9,8 @@ import openpnm.models.topology as tm
 logger = logging.getLogger(__name__)
 ws = Workspace()
 
+from openpnm.utils import tic, toc
+
 
 class GenericNetwork(Base, ModelsMixin):
     r"""
@@ -781,13 +783,14 @@ class GenericNetwork(Base, ModelsMixin):
         1
         """
         pores = self._parse_indices(pores)
-        # Count number of neighbors
-        num = self.find_neighbor_pores(pores, flatten=flatten,
-                                       mode=mode, include_input=True)
         if flatten:
+            # Count number of neighbors
+            num = self.find_neighbor_pores(pores, flatten=flatten,
+                                           mode=mode, include_input=True)
             num = np.size(num)
-        else:
-            num = np.array([np.size(i) for i in num], dtype=int)
+        else:   # Could be done much faster if flatten == False
+            am = self.create_adjacency_matrix(fmt="csr")
+            num = am[pores].sum(axis=1).squeeze()
         return num
 
     def find_nearby_pores(self, pores, r, flatten=False, include_input=False):