@ma-sadeghi 2020-07-09
@jgostick I ended up not using `numba` all. I tried a bunch of different approaches and none worked.